refactor float a double en todo

+++ src/pthreads.c
@@ -40,26 +40,26 @@ double tIni, tFin, tTotal;
 typedef struct cuerpo cuerpo_t;
 struct cuerpo
 {
-    float masa;
-    float px;
-    float py;
-    float pz;
-    float vx;
-    float vy;
-    float vz;
-    float r;
-    float g;
-    float b;
+    double masa;
+    double px;
+    double py;
+    double pz;
+    double vx;
+    double vy;
+    double vz;
+    double r;
+    double g;
+    double b;
     int cuerpo;
 };
 
-float *fuerza_totalX, *fuerza_totalY, *fuerza_totalZ;
-float toroide_alfa;
-float toroide_theta;
-float toroide_incremento;
-float toroide_lado;
-float toroide_r;
-float toroide_R;
+double *fuerza_totalX, *fuerza_totalY, *fuerza_totalZ;
+double toroide_alfa;
+double toroide_theta;
+double toroide_incremento;
+double toroide_lado;
+double toroide_r;
+double toroide_R;
 
 cuerpo_t *cuerpos;
 int delta_tiempo = 1.0f; // Intervalo de tiempo, longitud de un paso
@@ -67,13 +67,13 @@ int pasos;
 int N;
 int T;
 
-// void calcularFuerzas(int ini, int lim, float *local_fuerzaX, float *local_fuerzaY, float *local_fuerzaZ)
+// void calcularFuerzas(int ini, int lim, double *local_fuerzaX, double *local_fuerzaY, double *local_fuerzaZ)
 void calcularFuerzas(int idW)
 {
     int cuerpo1, cuerpo2;
-    float dif_X, dif_Y, dif_Z;
-    float distancia;
-    float F;
+    double dif_X, dif_Y, dif_Z;
+    double distancia;
+    double F;
 
     int slice = N / T;
     int ini = idW * slice;
@@ -329,9 +329,9 @@ int main(int argc, char const *argv[])
     T = atoi(argv[4]);
 
     cuerpos = (cuerpo_t *)malloc(sizeof(cuerpo_t) * N);
-    fuerza_totalX = (float *)malloc(sizeof(float) * N);
-    fuerza_totalY = (float *)malloc(sizeof(float) * N);
-    fuerza_totalZ = (float *)malloc(sizeof(float) * N);
+    fuerza_totalX = (double *)malloc(sizeof(double) * N);
+    fuerza_totalY = (double *)malloc(sizeof(double) * N);
+    fuerza_totalZ = (double *)malloc(sizeof(double) * N);
 
     inicializarCuerpos(cuerpos, N);
